Fix code style.

+++ src/gui/widgets/tabstrip.cpp
@@ -49,7 +49,7 @@ Widget *TabStrip::createWidget(const std::string &text) const
     widget->adjustSize();
     if (!mCount)
         widget->setPressed(true);
-    widget->setTag(mWidgets.size());
+    widget->setTag(static_cast<int>(mWidgets.size()));
     return widget;
 }
