+++ src/net/eathena/charserverhandler.cpp
@@ -569,8 +569,8 @@ void CharServerHandler::processCharRename(Net::MessageIn &msg)
 void CharServerHandler::changeSlot(const int oldSlot, const int newSlot)
 {
     createOutPacket(CMSG_CHAR_CHANGE_SLOT);
-    outMsg.writeInt16(oldSlot, "old slot");
-    outMsg.writeInt16(newSlot, "new slot");
+    outMsg.writeInt16(static_cast<int16_t>(oldSlot), "old slot");
+    outMsg.writeInt16(static_cast<int16_t>(newSlot), "new slot");
     outMsg.writeInt16(0, "unused");
 }
